Final Project - Paper Mario Day and Night
This is my Final Project for Drawing on the Web! For this project, I
created a scene inspired by Super Paper Mario. It includes elements from
various games in the series. The main function, in which you click the
floating man or floating Luigi, causes the scene to rotate, showing a
different perspective. This was a main feature of the gameplay in Super
Paper Mario.